revert libiconv update for 32-bit Windows
The replacement libiconv-2.dll was compiled with a newer MinGW than
before. Unfortunately, that changes the geenrated result in a way that
affects dependent libraries, such as libintl-9.dll. Recompiling
libintl-9.dll, in turn, creates a dependncy on libwinpthread-1.dll,
and the right solution there is probably to add libwinpthread-1.dll
--- but that's a kind of non-compatible change. So, defer this
update until we rebuild Windows libraries more generally.